### 2017-10-10 #422 @lheagy
Condensed matter physics isn't really my background, but I am happy to work with this submission. @arfon, is there anyone else one the editorial board that you think might have more background in this that could support?

@yxqd, I have a couple questions / comments regarding the paper:
- right now, the summary is brief and quite technical. I would recommend you expand it to make it more accessible for a broader audience. This could include some examples, or further discussion of applications.
- regarding the [notice of copyright](https://github.com/sns-chops/multiphonon/blob/master/paper/paper.md#notice-of-copyright) in the paper. All JOSS papers are licensed under a Creative Commons Attribution 4.0 License (http://joss.theoj.org/about#content_license). This does not appear to be a conflict, but I wanted to clarify that with you cc @arfon: are there any concerns with such a statement in the paper
- regarding testing, do you have coverage stats that could also be shown alongside the build status in the README? for example https://codecov.io/
