### 2017-10-10 #422 @arfon
> - regarding the notice of copyright in the paper. All JOSS papers are licensed under a Creative Commons Attribution 4.0 License (http://joss.theoj.org/about#content_license). This does not appear to be a conflict, but I wanted to clarify that with you cc @arfon: are there any concerns with such a statement in the paper

Thanks for the heads up @lheagy - I don't think this copyright statement causes us any major issues.
