### 2017-10-10 #422 @yxqd
@lheagy thanks! I admit I don't know enough about the editors backgrounds and did not find an editor specializing on condensed matter physics. You were the closest I could find from the editor list since studies on condensed matter in extreme conditions (high temperature and high pressure, for example) are topics of research in geophysics. Again, thanks @lheagy for picking this up.

@lheagy regarding your comments:

* I agree more explanations should help clarify the applications of the software. I will revise the summary as you suggested
* The [notice of copyright](https://github.com/sns-chops/multiphonon/blob/master/paper/paper.md#notice-of-copyright) was a standard text required by our lab. I don't think it is in conflict with Creative Commons. @arfon please let us know if there are problems
* Tests are run by travis CI right now. I will try to get coverage stats as you suggested.

I am travelling this week so I will only be able to work on these revisions next week. I will keep you posted. @lheagy @arfon please kindly update me on the editing and the copyright text. Thanks!
