### 2017-10-16 sns-chops/multiphonon#72 @yxqd
test coverage
Add test coverage using coverall.

Prompted by JOSS review openjournals/joss-reviews#422
