+++ .travis.yml
@@ -31,7 +31,7 @@ install:
   - conda config --set always_yes true
   - conda update conda
   - conda config --add channels conda-forge
-  - conda create -n testenv pip pytest python=$TRAVIS_PYTHON_VERSION
+  - conda create -n testenv pip pytest pytest-cov python=$TRAVIS_PYTHON_VERSION
   - conda install -n testenv numpy scipy matplotlib
   - conda install -n testenv -c mcvine histogram
   - conda install -n testenv -c neutrons ipywe
@@ -42,10 +42,13 @@ install:
 #  - pip install codecov
 
 script:
-   - py.test
+   - py.test --cov=multiphonon tests
    - travis_wait 40 python tests/getdos_TestCase.py
    - travis_wait 20 python tests/getdos_TestCase2.py
    - travis_wait 20 python tests/ui/batch_TestCase.py
+
+after_success:
+   - coveralls
 
 notifications:
   email:
