### 2017-10-16 6b9ca79
Refs #72. forgot to install coveralls
