+++ .travis.yml
@@ -38,14 +38,14 @@ install:
   - conda install -n testenv -c mantid mantid-framework
   - source activate testenv
   - python setup.py install
-#  - pip install coveralls
+  - pip install coveralls
 #  - pip install codecov
 
 script:
    - py.test --cov=multiphonon tests
-   - travis_wait 40 python tests/getdos_TestCase.py
-   - travis_wait 20 python tests/getdos_TestCase2.py
-   - travis_wait 20 python tests/ui/batch_TestCase.py
+#   - travis_wait 40 python tests/getdos_TestCase.py
+#   - travis_wait 20 python tests/getdos_TestCase2.py
+#   - travis_wait 20 python tests/ui/batch_TestCase.py
 
 after_success:
    - coveralls
