### 2017-10-16 e190d07
Refs #72. added sanity tests of UI routines. turned on a getdos test
