### 2017-10-16 045442d
Refs #72. need init value for context.sample_nxs to pass ui.getdos test
