### 2017-10-16 f495307
Refs #72. enable all tests in tests/getdos_TestCase.py
