### 2017-10-16 86efa59
Refs #72. enable getdos_TestCase2 and ui/batch_TestCase
