### 2017-10-17 f03f327
added coverall badge
