### 2017-10-17 7c462b1
trivial comment on travis config
