+++ .travis.yml
@@ -43,6 +43,7 @@ install:
 
 script:
    - travis_wait 60 py.test --cov=multiphonon tests
+# the following were run standalone but now all included in the cmd above.
 #   - travis_wait 40 python tests/getdos_TestCase.py
 #   - travis_wait 20 python tests/getdos_TestCase2.py
 #   - travis_wait 20 python tests/ui/batch_TestCase.py
