### 2017-10-17 sns-chops/multiphonon#74 @yxqd
Expand summary of JOSS paper
Refs openjournals/joss-reviews#422
