+++ paper/paper.bib
@@ -19,3 +19,60 @@
   doi = {10.1103/PhysRevB.75.104301},
   url = {https://link.aps.org/doi/10.1103/PhysRevB.75.104301}
 }
+
+@book{ashcroftmermin,
+  title={Solid State Physics},
+  author={Ashcroft, N.W. and Mermin, N.D.},
+  isbn={9788131500521},
+  year={2011},
+  publisher={Cengage Learning}
+}
+
+@article{lichen2015,
+  title={Orbitally driven giant phonon anharmonicity in SnSe},
+  author={Li, Chen W and Hong, Jiawang and May, Andrew F and Bansal, Dipanshu and Chi, Songxue and Hong, Tao and Ehlers, Georg and Delaire, O},
+  journal={Nature Physics},
+  volume={11},
+  number={12},
+  pages={1063},
+  year={2015},
+  publisher={Nature Publishing Group}
+}
+
+@article{budai2014,
+  title={Metallization of vanadium dioxide driven by large phonon entropy},
+  author={Budai, John D and Hong, Jiawang and Manley, Michael E and Specht, Eliot D and Li, Chen W and Tischler, Jonathan Z and Abernathy, Douglas L and Said, Ayman H and Leu, Bogdan M and Boatner, Lynn A and others},
+  journal={Nature},
+  volume={515},
+  number={7528},
+  pages={535},
+  year={2014},
+  publisher={Nature Publishing Group}
+}
+
+@article{FULTZ2010,
+title = "Vibrational thermodynamics of materials",
+journal = "Progress in Materials Science",
+volume = "55",
+number = "4",
+pages = "247 - 352",
+year = "2010",
+issn = "0079-6425",
+doi = "https://doi.org/10.1016/j.pmatsci.2009.05.002",
+url = "http://www.sciencedirect.com/science/article/pii/S0079642509000577",
+author = "Brent Fultz"
+}
+
+@phdthesis{swan2006vibrational,
+  title={Vibrational entropy contributions to the phase stability of iron-and aluminum-based binary alloys},
+  author={Swan-Wood, Tabitha Liana},
+  year={2006},
+  school={California Institute of Technology}
+}
+
+@phdthesis{bogdanoff2002phonon,
+  title={The phonon entropy of metals and alloys: the effects of thermal and chemical disorder},
+  author={Bogdanoff, Peter David},
+  year={2002},
+  school={California Institute of Technology}
+}
