### 2017-10-17 #422 @yxqd
@lheagy could you please check the expanded summary in [the revised paper](https://github.com/sns-chops/multiphonon/raw/da5a0097012df7679a8fc350f6949c9d095e0a3c/paper/paper.pdf)? The coverage badge were added to [the readme](https://github.com/sns-chops/multiphonon/blob/da5a0097012df7679a8fc350f6949c9d095e0a3c/README.md). Could you please find a reviewer for this paper? Thanks.
