### 2017-10-18 #422 @lheagy
👋 Hi @tacaswell and @bjmorgan, would you be willing to review this submission to JOSS?
