### 2017-10-21 #422 @bjmorgan
@lheagy Happy to review.
