### 2017-10-26 #422 @lheagy
@yxqd, sorry for the delay. It is preferable to have multiple reviewers, and is important here, as I really don't have much background in condensed matter physics.

I have sent emails to @tacaswell,  @betatim, @jochym (I couldn't find an email for @khinsen). If you know them personally and would also like to send an email, please feel free to point them to this issue. My email is lheagy@eos.ubc.ca if you would like to cc me.
