### 2017-10-27 #422 @jochym
@lheagy I'll be willing to review the paper. I do not have much of experience on experimental side but I have a background in phonon calculations - I am a theoretician. This would be my first paper at joss (I already know the http://joss.theoj.org/about#reviewer_guidelines ). What should be the next steps?
