### 2017-10-27 #422 @lheagy
Thanks @bjmorgan and @jochym for agreeing to review! I will close this issue and we will start a new issue that includes a checklist for the review. There is only one checklist, so please feel free to communicate back and forth on the issue if there are items to discuss and please don't hesitate to let me know if you have questions
