### 2017-10-27 #440 @whedon
[REVIEW]: Multiphonon: Phonon Density of States tools for Inelastic Neutron Scattering Powder Data
**Submitting author:** @Yxqd (<a href="http://orcid.org/0000-0001-9233-0100">Jiao Lin</a>)
**Repository:** <a href="https://github.com/sns-chops/multiphonon" target ="_blank">https://github.com/sns-chops/multiphonon</a>
**Version:** 0.1.1
