### 2017-10-27 #440 @lheagy
@bjmorgan, @jochym this is ready for you to review.

An informal guideline is that we would like your review in 2 weeks, but sooner (or later) are also ok.

Please carry out your review in this issue by updating the checklist above (please make sure you're logged in to GitHub). There is only one checklist, so please feel free to converse in this issue on the checklist items.

The reviewer guidelines are available here: http://joss.theoj.org/about#reviewer_guidelines

For detailed comments, please create a new issue in the [target repository](https://github.com/sns-chops/multiphonon), particularly if those are acceptance-blockers and link them to this thread (https://help.github.com/articles/autolinked-references-and-urls/)

Please let me know if you have any questions or concerns.

Thanks!!
