### 2017-10-27 #440 @jochym
Pardon my stupid question @lheagy , but where is the PDF of the paper? I can see the repo but I cannot locate proper text of the paper.
