### 2017-10-27 #440 @lheagy
Hi @jochym, the paper is here: https://github.com/sns-chops/multiphonon/tree/master/paper. There is a markdown, which renders nicely on github, and @yxqd also included a pdf in that same folder
