### 2017-10-27 #440 @jochym
Thanks. Did I miss something in guidelines? Is this supposed to be standard location?
