### 2017-10-27 #440 @lheagy
You are welcome! It is supposed to be in a standard location (in a paper folder at the root of the repository), however, we don't have this clearly documented in the reviewer guidelines or at the top of the issue. I have started an issue at openjournals/joss#351 so we can clarify that for reviewers
