### 2017-11-13 #440 @lheagy
👋 Hi @bjmorgan, @jochym, @tacaswell, I sent you all an email last week checking in. How are things going? Is there anything I can help with?

In terms of checking the boxes: @bjmorgan, would you be willing to take lead? @jochym and @tacaswell, can you give an overview of your thoughts on the status of each of the review items and if you have concerns or feel there are items that need to be improved prior to publication?

If you each prefer, I can give you individual check-lists rather than all filling in one. Please let me know if this would be more clear.

Thanks!
