### 2017-11-24 #440 @lheagy
👋 @bjmorgan, @jochym, @tacaswell: How are things going? Do you guys have an estimate of a timeline of when you will be able to take a look at this? Please let me know if I can be of assistance!
