### 2017-11-27 sns-chops/multiphonon#76 @bjmorgan
Submission to PyPI?
https://github.com/openjournals/joss-reviews/issues/440

The installation instructions currently require users to use `conda` for package management. Is it possible to submit this package to PyPI so that it can be installed using `pip`?
