### 2017-11-27 #440 @bjmorgan
https://github.com/sns-chops/multiphonon/issues/76
