### 2017-11-27 sns-chops/multiphonon#77 @bjmorgan
Test coverage
https://github.com/openjournals/joss-reviews/issues/440

Excellent to see this project using continuous integration, and reporting code coverage. Is there a reason why the tests do not cover a good proportion of the code (~25%)?
