### 2017-11-27 sns-chops/multiphonon#78 @bjmorgan
JOSS author list
https://github.com/openjournals/joss-reviews/issues/440

I am looking at the JOSS review criterion:
>Authorship: Has the submitting author (@yxqd) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

It seems unusual that people have contributed enough to be named in the first paragraph of the README, but not be included as authors on the JOSS paper. I wanted to check that this is deliberate?
