### 2017-11-27 sns-chops/multiphonon#79 @bjmorgan
Manually running tests
https://github.com/openjournals/joss-reviews/issues/440

Can instructions be added to the top level `README` for manually running the tests?
