### 2017-11-27 sns-chops/multiphonon#80 @bjmorgan
Community guidelines
https://github.com/openjournals/joss-reviews/issues/440

I can't find anything on the top level README that gives explicit information about these points:

> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support?
