### 2017-11-27 sns-chops/multiphonon#76 @yxqd
Thanks for the comments. Yes it is possible to create distutils-based distribution for this particular package but a couple dependencies of this package, including [histogram](https://github.com/danse-inelastic/histogram) and [mantid-framework](https://github.com/mantidproject/mantid/), contains c/c++ library and relies on cmake for building and releasing. conda is much easier an ecosystem for releases of multiphonon, so we chose not to build a PyPI package.
