### 2017-11-27 sns-chops/multiphonon#76 @bjmorgan
That’s reasonable. Would it be possible to add a comment in the installation section explaining this?

If the dependencies are already installed, can multiphonon be installed for source? e.g. `git clone …`, `python setup.py install`?
