### 2017-11-27 64f857f
added "community guideline"
