### 2017-11-27 sns-chops/multiphonon#78 @yxqd
Thank you @bjmorgan for your comments. The original multiphonon code was written more than 10 years ago by Max (the 2nd author of this JOSS paper) when he was writing his PhD thesis and several other authors made some improvements to it later on, fixing some annoyances. That code has shown some age and there are multiple versions with tweaks by various contributors, making it hard for users to choose which version to use. Besides, each version has its own problems. The new code is a response to the renewed interests in treating the powder INS data in a consistent, reproducible manner, and to the requirements gathered from users/contributors-of-the-original-code (see [the original whiteboard](https://github.com/sns-chops/multiphonon/blob/2c0cb427061af44e26a279c8008839ee9092265d/docs/initial_requirements.jpg) and [the github issue](https://github.com/sns-chops/multiphonon/issues/32)). We then decide that it is time to start from scratch and implement a new code. The code in this repository is therefore a complete rewrite  and the only code that resembles the previous code is in the [forward.phonon module](https://github.com/sns-chops/multiphonon/blob/64f857f75d9dd85633395a69383d301b3fef6378/multiphonon/forward/phonon.py), previously authored by Max. That being said, we are certainly grateful for explorations by and fruitful discussions with other contributors and they are acknowledged in the paper.
