### 2017-11-27 4446285
added a subsection explaining PyPI installation
Refs #76
