### 2017-11-27 sns-chops/multiphonon#76 @yxqd
Just added. And yes, `python setup.py install` works.
