### 2017-11-28 fb2a73b
Added instructions on running tests
