### 2017-11-28 sns-chops/multiphonon#79 @yxqd
Just added instructions. The tests can take a while to run. At travis-CI, it recently took about 20~30 minutes per build. BTW, instructions for running examples are [here](https://github.com/sns-chops/multiphonon/tree/master/examples)
