### 2017-11-28 sns-chops/multiphonon#77 @yxqd
Thanks, @bjmorgan. One module that is not yet tested at all contains some helper functions for plotting: https://coveralls.io/builds/14390180/source?filename=multiphonon%2Fbackward%2Fplotutils.py. But I agree some modules deserve more coverage and I will work on them.
