### 2017-11-28 #440 @bjmorgan
@jochym @tacaswell I'm a solid state computational theorist and only deal with things like phonon densities of states *after* they have been extracted from experimental data. Are either of you two in a position to check whether the inputs / outputs for the examples look how you would expect?
