### 2017-11-28 c264c81
Refs #77. trying to add a test of plotutils.plot_residual
