### 2017-11-28 1f5a054
Refs #77. trying to make plot_residual test work under travis CI
