+++ tests/backward/plotutils_TestCase.py
@@ -4,7 +4,6 @@
 interactive = False
 
 import sys, os, warnings
-from multiphonon.backward import plotutils
 datadir = os.path.join(os.path.dirname(__file__), "../data")
 workdir = os.path.join(datadir, 'work-V')
 here = os.path.dirname(__file__)
@@ -14,10 +13,20 @@ import unittest
 class TestCase(unittest.TestCase):
 
 
+    def setUp(self):
+        if not interactive:
+            import matplotlib
+            matplotlib.use('Agg')
+        from multiphonon.backward import plotutils
+        from matplotlib import pyplot as plt
+        self.pu = plotutils
+        self.plt = plt
+
+
     def test1(self):
-        plotutils.plot_residual(workdir)
+        self.pu.plot_residual(workdir)
         if interactive:
-            plt.show()
+            self.plt.show()
         return
         
         
@@ -25,9 +34,7 @@ class TestCase(unittest.TestCase):
 
 
 if __name__ == "__main__":
-    global interactive
     interactive = True
-    from matplotlib import pyplot as plt
     unittest.main()
     
 # End of file 
