### 2017-11-28 4d44740
Refs #77. more tests for plotutils
