+++ tests/backward/plotutils_TestCase.py
@@ -24,7 +24,32 @@ class TestCase(unittest.TestCase):
 
 
     def test1(self):
+        "plot_residual"
         self.pu.plot_residual(workdir)
+        if interactive:
+            self.plt.show()
+        return
+        
+        
+    def test2(self):
+        "plot_dos_iteration"
+        self.pu.plot_dos_iteration(workdir)
+        if interactive:
+            self.plt.show()
+        return
+        
+        
+    def test3(self):
+        "plot_intermediate_result_sqe"
+        self.pu.plot_intermediate_result_sqe(os.path.join(workdir, 'round-5'))
+        if interactive:
+            self.plt.show()
+        return
+        
+        
+    def test4(self):
+        "plot_intermediate_result_se"
+        self.pu.plot_intermediate_result_se(os.path.join(workdir, 'round-5'))
         if interactive:
             self.plt.show()
         return
