### 2017-11-28 cd3848f
Refs #77. hide some obsolete code in sqe2dos
