### 2017-11-28 9c4e86a
Refs #77. test dos.nice.smooth
