### 2017-11-28 1f1fa7c
Refs #77. test sqe.plot
