### 2017-11-28 f3f3d82
Refs #77. hide obsolete code in forward.phonon
