### 2017-11-28 095187e
Refs #77. backward/singlephonon_sqe2dos.py: hide obsolete code
