### 2017-11-28 sns-chops/multiphonon#77 @yxqd
Hi @bjmorgan, test coverage was improved by #81 to 88%. Please check. Thanks.
