### 2017-12-14 #440 @jochym
@lheagy: Thanks for not giving up on me with this review. Here is my opinion on this submission (based on the checklist above):

## Conflict of interest
As the reviewer I confirm that I have read the JOSS conflict of interest policy and that there are no conflicts of interest for me to review this work.

## Code of Conduct
 I confirm that I read and will adhere to the JOSS code of conduct.

## General checks
[Y] Repository: Is the source code for this software available at the repository url?
[Y] License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
[Y] Version: Does the release version given match the GitHub release (0.1.1)?
[Y] Authorship: Has the submitting author (@yxqd) made major contributions to the software?
[Y] Does the full list of paper authors seem appropriate and complete?

**Note**: _The last point is difficult to judge. The development was clearly carried out outside of github, and only fairly recent changes are tracked. I have no reason to doubt the declarations of the submitter. But I have also no real way to verify them._

## Functionality
### Installation: Does installation proceed as outlined in the documentation?

**Essentially yes**, but I have few problems with the procedure:
- The list of dependencies is _huge_. It may be fully justified but from the docs (just two functions in the API) it seems that the dependency management is too careless. Furthermore it depends on packages from four (!) different conda channels. This install is bound to have problems quite soon due to the version drift and incompatibilities - see below.
- The package is written in the heading into obsolescence python 2.7 (EOL in 2020). The process of adoption of python 3 is speeding up quite a bit recently. Multiple new developments in data science already depend on python 3. Including technologies used by the authors - parts of jupyter are already python 3 only. This cannot be a reason for rejection but I would encourage authors to port to python 3  soon - it is easier than expected.
- Such complicated install forces the package into separate virtual environment and makes it difficult or impossible to integrate the package into larger production environment. E.g. it would be impossible to integrate the package into standard jupyterhub environment without major work and inconveniences. This substantially decreases the utility of the software.
- Non-specified dependency on `pyversion` makes 5 tests to fail. Furthermore this dependency cannot be installed from conda, it's last release on pypi was in 2015 and does not support python 3. Essentially package cannot be fully tested using described install procedure.

### Functionality: Have the functional claims of the software been confirmed?
**No**.
- I was able to run command line example, albeit not cleanly. I have got a following warning:
```
python2.7/site-packages/histogram/hdf/Loader.py:129: FutureWarning: elementwise comparison failed; returning scalar instead, but in the future will perform elementwise comparison
  if 'storage' in list(dataGroup): # this uses the 'storage' convention
```
which does not increase confidence in the code - suggesting that it will soon become obsolete.
- Some (5) tests included in the package fail due to the missing pyversion dependence and there are two warnings. One listed above and the second hard to interpret:
```
tests/backward/singlephonon_sqe2dos_TestCase.py::TestCase::test1c
  /home/jochym/Projects/joss-review/multiphonon/multiphonon/backward/singlephonon_sqe2dos.py:125: UserWarning: Scaling factor to combine DOSes calculated is not stable: 22.1354562834 (using continuous criteria) vs 17.7865062045 (using area criteria)
  You may want to consider adjusting the parameters such as E range (Emax more specifically)
    "You may want to consider adjusting the parameters such as E range (Emax more specifically)" % (scale1, scale2)
```
- The example notebook cannot be executed to the end. The `Converting sample data to powder I(Q,E)...` stage produced stack dump with following import error:
```
ImportError: libjsoncpp.so.0: cannot open shared object file: No such file or directory
```
probably due to another missing or changed dependency.
Since the notebook is the only fully worked-out example I cannot say the functionality was verified. In fact it will be difficult to even start to fully use the software - having only short, simple and tersely commented example `getdos-Al.py` to start with.
### Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
**There are no particular performance claims**. The performance during working tests/examples was fully satisfactory.

## Documentation

### A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
**Yes**
### Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
**Mostly** - the install is a bit complicated but described in enough detail to be followed and completed on the first try. Some dependencies are mentioned, but see my comments above. Some dependencies are missing - non-essential and apparently also essential (see above).

### Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
**Yes**, but only the simplest example run to completion and the more elaborated (notebook) failed to execute (see above).
### Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
**No**. There is only very limited API documentation for two functions. Judging from the simplest example, even this simplest usage requires calling additional API functions (`multiphonon.sqe.interp`). The only functional documentation is a tersely commented notebook which I was unable to run (see above).

**This level of documentation makes the software not suitable for publication**

### Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?
**Yes**. But some of them fail.

### Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
**Yes**

## Software paper
[Y] Authors: Does the paper.md file include a list of authors with their affiliations?
[Y] A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
[N] References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
**Note**: Some references have missing doi (e.g. both Nature papers)

## Conclusion
Due to the above mentioned deficiencies, in particular:
- Missing functional documentation
- Non-functional examples
- Other smaller deficiencies mentioned above

I cannot recommend publication of this submission at this state.
However, I believe that the mentioned problems can be resolved, which may warrant reconsideration.
