### 2017-12-14 #440 @jochym
@bjmorgan Unfortunately I am theoretician myself - thus I cannot check the inputs but the outputs seem at least reasonable (i.e. when I could get them out - see my report).
