### 2017-12-15 sns-chops/multiphonon#82 @lheagy
JOSS Review: Installation 
Originally from @jochym in openjournals/joss-reviews#440

- The list of dependencies is _huge_. It may be fully justified but from the docs (just two functions in the API) it seems that the dependency management is too careless. Furthermore it depends on packages from four (!) different conda channels. This install is bound to have problems quite soon due to the version drift and incompatibilities - see below.
- The package is written in the heading into obsolescence python 2.7 (EOL in 2020). The process of adoption of python 3 is speeding up quite a bit recently. Multiple new developments in data science already depend on python 3. Including technologies used by the authors - parts of jupyter are already python 3 only. This cannot be a reason for rejection but I would encourage authors to port to python 3  soon - it is easier than expected.
- Such complicated install forces the package into separate virtual environment and makes it difficult or impossible to integrate the package into larger production environment. E.g. it would be impossible to integrate the package into standard jupyterhub environment without major work and inconveniences. This substantially decreases the utility of the software.
- Non-specified dependency on `pyversion` makes 5 tests to fail. Furthermore this dependency cannot be installed from conda, it's last release on pypi was in 2015 and does not support python 3. Essentially package cannot be fully tested using described install procedure.

### Items to be addressed
- [ ] dependencies: is the large dependency list necessary? If so, could this be explained in the documentation, otherwise, removing dependencies that are not necessary will simplify installation and maintenance
- [ ] is there a plan to upgrade to python 3? This does not need to be completed prior to an accepted publication, but explaining if this is in the roadmap in the documentation would be helpful
- [ ] can the installation process be simplified? if not, please provide some explanation in the documentation
- [ ] `pyversion` dependency and failing tests
