### 2017-12-15 sns-chops/multiphonon#83 @lheagy
JOSS Review: Functionality and Examples
Originally from @jochym in openjournals/joss-reviews#440

- I was able to run command line example, albeit not cleanly. I have got a following warning:
```
python2.7/site-packages/histogram/hdf/Loader.py:129: FutureWarning: elementwise comparison failed; returning scalar instead, but in the future will perform elementwise comparison
  if 'storage' in list(dataGroup): # this uses the 'storage' convention
```
which does not increase confidence in the code - suggesting that it will soon become obsolete.
- Some (5) tests included in the package fail due to the missing pyversion dependence and there are two warnings. One listed above and the second hard to interpret:
```
tests/backward/singlephonon_sqe2dos_TestCase.py::TestCase::test1c
  /home/jochym/Projects/joss-review/multiphonon/multiphonon/backward/singlephonon_sqe2dos.py:125: UserWarning: Scaling factor to combine DOSes calculated is not stable: 22.1354562834 (using continuous criteria) vs 17.7865062045 (using area criteria)
  You may want to consider adjusting the parameters such as E range (Emax more specifically)
    "You may want to consider adjusting the parameters such as E range (Emax more specifically)" % (scale1, scale2)
```
- The example notebook cannot be executed to the end. The `Converting sample data to powder I(Q,E)...` stage produced stack dump with following import error:
```
ImportError: libjsoncpp.so.0: cannot open shared object file: No such file or directory
```
probably due to another missing or changed dependency.
Since the notebook is the only fully worked-out example I cannot say the functionality was verified. In fact it will be difficult to even start to fully use the software - having only short, simple and tersely commented example `getdos-Al.py` to start with.
