### 2017-12-15 sns-chops/multiphonon#84 @lheagy
JOSS Review: Documentation
Originally from @jochym in openjournals/joss-reviews#440

There is only very limited API documentation for two functions. Judging from the simplest example, even this simplest usage requires calling additional API functions (`multiphonon.sqe.interp`). The only functional documentation is a tersely commented notebook which I was unable to run (see above).
