### 2017-12-15 sns-chops/multiphonon#85 @lheagy
JOSS Review: Paper doi's
Originally from @jochym in openjournals/joss-reviews#440

Some references have missing doi (e.g. both Nature papers)
