### 2017-12-15 #440 @lheagy
Thanks for your comments @jochym.

@yxqd: could you please take a look at @jochym's comments above? These items are resolvable, and I have started issues where they can be tracked (@jochym: please add to or comment on these as you see fit).

- sns-chops/multiphonon#82
- sns-chops/multiphonon#83
- sns-chops/multiphonon#84
- sns-chops/multiphonon#85

Once these are addressed, we can continue the review process. Please let me know if there is anything I can clarify
