### 2017-12-15 #440 @yxqd
Thank you, @jochym and @lheagy !  I will take actions on these tickets. Some are relatively quick to answer, but some probably won't all be resolved immediately. Your patience is much appreciated!
