### 2017-12-15 sns-chops/multiphonon#83 @yxqd
@jochym could you please kindly let us know in detail at what kind of system (linux distribution, conda version, etc) your running of tests failed? We would like to reproduce the error and solve the installation problem.
