### 2017-12-15 sns-chops/multiphonon#82 @yxqd
**Dependencies**

This deserves a bit of explanation. The fact that the current dependency is long is probably mostly due to it now depends on [mantid](http://www.mantidproject.org/), which is a huge project and depends on [a lot of packages](https://github.com/mantidproject/conda-recipes/blob/master/framework/meta.yaml). The core functionality of the multiphonon package (converting from I(Q,E) spectrum to DOS spectrum) does not require Mantid, as you can see in [this simple examle](https://github.com/sns-chops/multiphonon/blob/61dd792aa3695281240abf51e9151a308e86163c/examples/getdos-Al.py). The Mantid package is used solely for reducing the raw data from the inelastic neutron scattering experiment to I(Q,E) spectrum. So we could remove the mantid dependency to make this package minimal. The reason we had Mantid in the dependency list was for practical usage of the multiphonon package. This software, deployed at the [SNS](https://neutrons.ornl.gov/sns) jupyter hub environment, is for users of the direct geometry inelastic neutron scattering instruments, and the Mantid software is the standard solution for reducing the raw data at these instruments. It is almost inevitable for users of this package to be using mantid at some stage, at least for now. However, I agree that this binding of the multiphonon package and the mantid software can happen in some other way. So I plan to

* [ ] remove Mantid from the dependency list, and
* [ ] revise installation instructions to let users know how to install Mantid if needed
