### 2017-12-15 sns-chops/multiphonon#82 @yxqd
@jochym @lheagy please comment if this looks reasonable to you. Thanks!
