### 2017-12-15 sns-chops/multiphonon#83 @yxqd
See also #82
