### 2017-12-15 sns-chops/multiphonon#84 @yxqd
List of functions to add API doc or to expose API doc:

* https://github.com/sns-chops/multiphonon/blob/61dd792aa3695281240abf51e9151a308e86163c/multiphonon/redutils.py#L1
* https://github.com/sns-chops/multiphonon/blob/61dd792aa3695281240abf51e9151a308e86163c/multiphonon/backward/singlephonon_sqe2dos.py#L10
* Plotting methods in https://github.com/sns-chops/multiphonon/blob/61dd792aa3695281240abf51e9151a308e86163c/multiphonon/backward/plotutils.py
* https://github.com/sns-chops/multiphonon/blob/61dd792aa3695281240abf51e9151a308e86163c/multiphonon/forward/__init__.py#L7
* https://github.com/sns-chops/multiphonon/blob/61dd792aa3695281240abf51e9151a308e86163c/multiphonon/forward/phonon.py#L21
* https://github.com/sns-chops/multiphonon/blob/61dd792aa3695281240abf51e9151a308e86163c/multiphonon/forward/phonon.py#L30
* https://github.com/sns-chops/multiphonon/blob/61dd792aa3695281240abf51e9151a308e86163c/multiphonon/sqe/__init__.py#L27
