### 2017-12-15 sns-chops/multiphonon#82 @jochym
## pyversion.
I have created the install in fairly standard way:
- `conda create -n review -c conda-forge python=2.7`
- `source activate review`
- `conda install -c ....... multiphonon`

Then after cloning of the distro I have tried to run the tests as described in the docs. Five tests failed due to the pyversion dependency. This was run on my debian testing (buster) machine. I can try the same in the clean debian stretch virtual machine but probably in few days.

## dependencies
I understand many users of multiphonon will also use mantid but this is not a good reason to make it a dependency if it really does not _depend_ on it for its functionality. The dependency list should be as short as possible in my opinion. Let the users decide what they need. You can always create simple meta-package (or even better conda environment) for your local setup which just depends on popular packages.

## python 3
As I stated in my review this is not a show-stopper, but I predict that the problems with increasing obsolescence of py2 packages will soon be substantial.

## Installation
It would help to either make a copy of packages from other channels or, better still, package missing parts for conda-forge. Again, I do not consider this a show-stopper.
