### 2017-12-15 sns-chops/multiphonon#83 @jochym
The test system was my working machine. Debian buster with the newest conda on python 3.6 - freshly updated. All tests were run in separate environment created just for the review according to the docs.
I am using the same system for all my work and development - which is mostly in conda environments.
I'll try to help you by sharing the conda environment file: https://anaconda.org/jochym/review
I can also try to repeat my tests on freshly installed virtual machine - but that will have to wait some time.
