### 2017-12-15 sns-chops/multiphonon#84 @jochym
## API
- Try to aim for the level of documentation used in scipy - such that users will have all necessary data to use your code.
- Please add some overview paragraphs to the docs to help users orient themselves.

## Functional docs
The notebook tutorial is a very good idea and it is mostly fine (except it did not run on my system). I have only one comment regarding the contents: please consider adding more narrative material, some overview and step by step explanation.  Remember you are writing for a novice user not an expert (they can figure out everything without your help).
