### 2017-12-17 a7f9cbc
Refs #86. updated instructions and conda-recipe
