### 2017-12-17 7f771af
Refs #86. added instruction to run tests without mantid
