+++ README.md
@@ -109,7 +109,11 @@ See http://sns-chops.github.io/multiphonon/
 ## Run tests
 Tests are run automatically at [travis CI](https://travis-ci.org/sns-chops/multiphonon/builds). 
 
-To manually run tests, install multiphonon, clone this repository, and cd into the new directory, and then run
+To manually run tests that do not depend on mantid, install multiphonon, clone this repository, and cd into the new directory, and then run
+
+    $ py.test -m "not needs_mantid"
+
+If mantid was installed, you can run all tests by
 
     $ py.test
 
