### 2017-12-17 1c595a6
slight refactor of multiphonon.ui: moved "Context" to toplevel as it is used by both batch and getdos
