### 2017-12-17 2b279ed
Refs #88. moving to 0.1.2a1
