+++ README.md
@@ -31,7 +31,6 @@ The multiphonon package can be installed using conda (python 2) on a recent 64bi
 
       $ conda config --add channels conda-forge
       $ conda config --add channels mcvine
-      $ conda install numpy
       $ conda install multiphonon
 
 Information on dependencies of this code can be found at [the conda recipe](/conda-recipe/meta.yaml)
