+++ conda-recipe/meta.yaml
@@ -1,6 +1,6 @@
 package:
   name: multiphonon
-  version: "0.1.1"
+  version: "0.1.2a1"
 
 source:
   git_rev: master
@@ -20,6 +20,7 @@ requirements:
     - scipy
     - matplotlib
     - pytest
+    - pyyaml
 #    - mantid-framework
 #    - ipywe
 
