### 2017-12-17 a8739a5
mark tests that need ipywe. Refs #86
