+++ pytest.ini
@@ -1,4 +1,0 @@
-[pytest]
-python_files=*TestCase*.py
-markers =
-  needs_mantid: maker a test that needs mantid
