+++ tests/pytest.ini
@@ -1,0 +1,5 @@
+[pytest]
+python_files=*TestCase*.py
+markers =
+  needs_mantid: mark a test that needs mantid
+  needs_ipywe: mark a test that needs ipywe
