### 2017-12-17 fe4a400
added DOIs and URLs. Refs #85
