+++ paper/paper.bib
@@ -36,7 +36,8 @@
   number={12},
   pages={1063},
   year={2015},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/nphys3492}
 }
 
 @article{budai2014,
@@ -47,7 +48,8 @@
   number={7528},
   pages={535},
   year={2014},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/nature13865}
 }
 
 @article{FULTZ2010,
@@ -67,12 +69,14 @@ author = "Brent Fultz"
   title={Vibrational entropy contributions to the phase stability of iron-and aluminum-based binary alloys},
   author={Swan-Wood, Tabitha Liana},
   year={2006},
-  school={California Institute of Technology}
+  school={California Institute of Technology},
+  url={http://resolver.caltech.edu/CaltechETD:etd-09012005-143247}
 }
 
 @phdthesis{bogdanoff2002phonon,
   title={The phonon entropy of metals and alloys: the effects of thermal and chemical disorder},
   author={Bogdanoff, Peter David},
   year={2002},
-  school={California Institute of Technology}
+  school={California Institute of Technology},
+  url={http://resolver.caltech.edu/CaltechETD:etd-12272005-154952}
 }
