### 2017-12-18 78f804c
Refs #89. added a method in redutils module
