+++ multiphonon/redutils.py
@@ -1,3 +1,18 @@
+def _createDefaultMantidUserConfig(facility='SNS'):
+    # create default Mantid user configuration for DEMO purpose.
+    import os
+    mantid_config_path = os.path.expanduser('~/.mantid/Mantid.user.properties')
+    mantid_user_dir = os.path.dirname(mantid_config_path)
+    if not os.path.exists(mantid_config_path):
+        if not os.path.exists(mantid_user_dir):
+            os.makedirs(mantid_user_dir)
+            with open(mantid_config_path, 'wt') as of:
+                of.write('default.facility=%s' % facility)
+    return
+# this should be done before mantid is imported
+_createMantidUserConfig()
+
+
 def reduce(nxsfile, qaxis, outfile, use_ei_guess=False, ei_guess=None, eaxis=None, tof2E=True, ibnorm='ByCurrent'):
     from mantid.simpleapi import DgsReduction, SofQW3, SaveNexus, Load
     if tof2E == 'guess':
@@ -59,3 +74,5 @@ def extract_iqe(mantid_nxs, histogram):
     import histogram.hdf as hh
     hh.dump(iqe, outpath)
     return
+
+
