### 2017-12-18 3362dee
Refs #91. start working on redutils.reduce
