### 2017-12-18 ccb0b59
Refs #89. fixed typo
