### 2017-12-18 30b80a9
Refs #89. fixed a bug
