### 2017-12-18 3b5fcde
Revert "Improve API doc"
