### 2017-12-18 074ae24
Refs #91. improving reduitls.reduce
