+++ multiphonon/redutils.py
@@ -1,5 +1,33 @@
 def reduce(nxsfile, qaxis, outfile, use_ei_guess=False, ei_guess=None, eaxis=None, tof2E=True, ibnorm='ByCurrent'):
-    "reduce Nexus file to I(Q,E)"
+    """reduce Nexus file to I(Q,E)
+
+    Parameters
+    ----------
+
+    nxsfile: str
+        path to nxs file
+
+    qaxis: float
+	Momentum transfer axis
+
+    outfile: str
+	path to save nxs data
+
+    use_ei_guess:float
+	Use of incident energy guessed (meV)
+
+    ei_guess: float
+	Initial guess of incident energy (meV) 
+
+    eaxis: float
+	Energy transfer axis
+
+    tof2E: float
+	Conversion from time of flight axis to energy axis
+
+    ibnorm: float
+	Incident beam normalization factor 
+    """
     from mantid.simpleapi import DgsReduction, SofQW3, SaveNexus, Load
     if tof2E == 'guess':
         # XXX: this is a simple guess. all raw data files seem to have root "entry"
