### 2017-12-18 4691a1f
add explanation of parameters
