+++ multiphonon/forward/__init__.py
@@ -8,6 +8,28 @@ def dos2sqe(dos, C_ms, sqe, T, M, Ei):
     """calculate sqe from DOS.
 
     The computed sqe has similar props as the given (experimental) sqe.
+
+    Parameters
+    ----------
+    dos: histogram
+	Phonon density of states
+
+    C_ms:float
+	MS = C_ms * MP
+
+    sqe:histogram
+	S(Q,E)
+
+    T: float
+	Temperature (Kelvin)
+    
+    M: float
+	Atomic mass
+    
+    Ei: float
+	Incident energy (meV)
+  
+
     """
     Q, E = sqe.Q, sqe.E
     dQ = Q[1]-Q[0]; Qmin = Q[0]; Qmax=Q[-1] + dQ/2.
