### 2017-12-18 cf21943
add explanation of parameters
