### 2017-12-18 476c187
multiphonon/redutils.py: added a step to check Mantid
