### 2017-12-18 b026e45
multiphonon/redutils: trivial
