### 2017-12-18 100e315
add explanation of parameters
