+++ multiphonon/backward/plotutils.py
@@ -4,6 +4,16 @@ import histogram.hdf as hh, os
 
 
 def plot_dos_iteration(curdir, total_rounds=None):
+    """plot the  DOS for each iteration
+   
+    Parameters
+    ----------
+   curdir: str
+	path to the current directory
+ 
+   total_rounds: integer
+	number of iterations
+   """
     if total_rounds is None:
         import glob
         dirs = glob.glob(os.path.join(curdir, 'round-*'))
@@ -20,6 +30,14 @@ def plot_dos_iteration(curdir, total_rounds=None):
 
 
 def plot_residual(curdir):
+   """plot the  residual DOS
+   
+    Parameters
+    ----------
+   curdir: str
+        path to the current directory
+   """
+
     exp_pos_se = hh.load(os.path.join(curdir, 'I_E-exp-posE.h5'))
     residual_pos_se = hh.load(os.path.join(curdir, 'residual_E-posE.h5'))
     E = exp_pos_se.E
@@ -31,6 +49,14 @@ def plot_residual(curdir):
 
 
 def plot_intermediate_result_sqe(curdir):
+     """plot the  intermediate S(Q,E)
+   
+    Parameters
+    ----------
+   curdir: str
+        path to the current directory
+   """
+
     from ._sqe2dos_script_templates import plots_table as plots
     plots = plots.strip().splitlines()
     plots = [p.split() for p in plots]
@@ -58,6 +84,14 @@ def plot_intermediate_result_sqe(curdir):
 
 
 def plot_intermediate_result_se(curdir):
+    """plot the  intermediate S(E)
+   
+    Parameters
+    ----------
+   curdir: str
+        path to the current directory
+   """
+
     # mpl.rcParams['figure.figsize'] = 12,9
     from ._sqe2dos_script_templates import plots_table as plots
     plots = plots.strip().splitlines()
