### 2017-12-19 84cdca7
add explanation of parameters
