+++ multiphonon/sqe/__init__.py
@@ -6,6 +6,13 @@
 import histogram as H, numpy as np
 
 def plot(iqe):
+ """ Parameters
+    ----------
+
+    iqe: histogram
+        input IQE
+
+ """
     Q = iqe.Q
     try:
         E= iqe.energy
@@ -27,8 +34,15 @@ def plot(iqe):
 def interp(iqehist, newE):
     """compute a new IQE histogram using the new energy array
     
-    * iqehist: input IQE histogram
-    * newE: new energy centers array
+    Parameters     
+    ----------
+
+    iqehist: histogram
+	input IQE
+
+    newE:float 
+	new energy centers array
+
     """
     from scipy import interpolate
     mask = iqehist.I != iqehist.I
@@ -87,6 +101,16 @@ def dynamical_range_mask(sqe, Ei):
     """calculate a mask of dynamical range being measured
     at the given incident energy.
     0 means within dynamical range
+
+    Parameters     
+    ----------
+
+    sqe: histogram
+        S(Q,E)
+
+    Ei:float 
+        incident energy
+
     """
     Q = sqe.Q
     E = sqe.E
