### 2017-12-19 14dbeba
Refs #84. added more details to examples/getdos2-V_Ei120meV.ipynb
