### 2017-12-19 7ca0bd5
add explanation of parameters
