+++ multiphonon/forward/phonon.py
@@ -161,11 +161,34 @@ def iterSQESet(N, Q,dQ, E,dE, M, g, beta):
 
 def computeSQESet(N, Q,dQ, E,dE, M, g, beta):
     """compute the set of S(Q,E) for n in [1,N]
-    Q, dQ: Q axis
-    E, dE: E axis
-    M: mass
-    g: phonon DOS for the given E
-    beta: 1/(kBT)
+
+    Parameters
+    ----------
+
+    N:integer
+         number of iterations
+
+    Q:float
+	momentum transfer axis
+
+    dQ:float
+	step size for momentum transfer axis
+ 
+    E: float
+	energy transfer axis
+
+    dE:float
+	step size for energy transfer axis
+ 
+    M:float
+	atomic  mass
+
+    g:float
+	 phonon DOS for the given E
+
+    beta:float
+	 1/(kBT)
+    
     """
     
     E2, AnE_set = computeAnESet(N, E,g, beta, dE)
@@ -183,6 +206,16 @@ def computeSQESet(N, Q,dQ, E,dE, M, g, beta):
 
 def computeSnQSet(N, DW2):
     """ compute the set of Sn(Q) for n in [1,N]
+
+    Parameters
+    ----------
+
+    N:integer
+         number of iterations
+
+    DW2: float
+	Debye Waller factor
+
     """
     SNQ = []
     for i in range(1,N+1):
@@ -194,12 +227,41 @@ def computeSNQ(DW2,N):
     """ Takes the exponent for the Debye Waller factor `DW2` = 2W, and an
     integer N indicating a term in the phonon expansion and returns the 
     intensity of the N-phonon incoherent scattering S_N(Q)
+
+    Parameters
+    ----------
+
+    N:integer
+         a term in the phonon expansion
+
+    DW2: float
+        Debye Waller factor
+
     """
     return DW2**N * np.exp(-DW2) / float(math.factorial(N))
 
 
 def computeAnESet(N, E,g, beta, dE):
     """compute the set of An(E) for n in [1,N]
+
+    Parameters
+    ----------
+
+    N:integer
+         number of iterations
+ 
+    E: float
+	energy transfer axis
+
+    dE:float
+	step size for energy transfer axis
+
+    g:float
+	 phonon DOS for the given E
+
+    beta:float
+	 1/(kBT)
+
     """
     E, A1E = computeA1E(E,g, beta, dE)
     ANE = np.zeros((N,) + A1E.shape, dtype=A1E.dtype)
@@ -215,6 +277,19 @@ def AnE_from_n_1(A1E, Anm1E, dE):
     """compute A_n(E) from A_{n-1}(E)
 
     A_n(E) = A1 (convolve) A_{n-1}
+
+    Parameters
+    ----------
+
+    A1E:float
+         array of energies
+
+   Anm1E:float
+	array of previous energies
+
+   dE:float
+	step size for energy transfer axis
+
     """
     Y = np.zeros( 4*len(Anm1E),'d' )
     Y[len(A1E):2*len(A1E)] = Anm1E
@@ -233,7 +308,15 @@ def AnE_from_n_1(A1E, Anm1E, dE):
 
 
 def convMatrix(y):
-    """ Returns matrix M, whose rows are filled with shifted copies of vector y"""
+    """ Returns matrix M, whose rows are filled with shifted copies of vector y
+
+    Parameters
+    ----------
+
+    y:float  
+         a vector
+
+    """
     M = np.zeros( ( len(y),len(y) ) , 'd' )
     for i in range(len(y)):
         M[i,i:] = y[:len(y)-i]
@@ -245,14 +328,25 @@ def computeA1E(E,g, beta, dE):
     
     A_1(E) = g(E)/(E*gamma_0) / (exp(E/kBT) - 1)
 
[…]
