+++ multiphonon/backward/singlephonon_sqe2dos.py
@@ -4,35 +4,36 @@
 
 import numpy as np, histogram as H, histogram.hdf as hh, os
 
+
 class EnergyAxisMissingBinCenterAtZero(Exception): pass
 
 
 def sqe2dos(sqe, T, Ecutoff, elastic_E_cutoff, M, initdos=None, update_weights=None):
-    """ 
+    """
     Given a one-phonon sqe, compute dos
 
     Parameters
     ----------
     sqe:histogram
-	S(Q,E)
+        S(Q,E)
 
     T:float
       Temperature (kelvin)
 
-    Ecutoff:float 
-	Cutoff energy beyond which DOS must be zero
+    Ecutoff:float
+        Cutoff energy beyond which DOS must be zero
     
-    Elastic_E_cutoff:float 
-	Cutoff energy for removing the elastic line
+    Elastic_E_cutoff:float
+        Cutoff energy for removing the elastic line
     
-    M:float 
-	Atomic mass
+    M:float
+        Atomic mass
     
     initdos:histogram
-	 initial guess of DOS
+        initial guess of DOS
 
-    update_weights:float 
-	weights for DOS update strategies (continuity, area conservation)
+    update_weights:float
+        weights for DOS update strategies (continuity, area conservation)
 
     The basic procedure is
     * construct an initial guess of DOS
@@ -40,13 +41,13 @@ def sqe2dos(sqe, T, Ecutoff, elastic_E_cutoff, M, initdos=None, update_weights=N
     * for both exp and sim SQE, integrate along Q to obtain S(E)
     * scale the initial guess DOS by the S(E) ratio
     * optionally we can do this again
-    """ 
+    """
     # create initial guess of dos
     Efull = sqe.E
     dE = sqe.E[1] - sqe.E[0]
-    assert dE>0, "Energy axis must be incremental"
-    Eplus = Efull[Efull>-dE/2]
-    if abs(Eplus[0]) > dE/1e6:
+    assert dE > 0, "Energy axis must be incremental"
+    Eplus = Efull[Efull > -dE / 2]
+    if abs(Eplus[0]) > dE / 1e6:
         raise EnergyAxisMissingBinCenterAtZero('"0" must be one of the bin centers of the energy axis')
     Eplus[0] = 0.
     if initdos is None:
@@ -63,14 +64,14 @@ def sqe2dos(sqe, T, Ecutoff, elastic_E_cutoff, M, initdos=None, update_weights=N
     dQ = Q[1] - Q[0]
     E = sqe.E
     dE = E[1] - E[0]
-    beta = 1./(T*kelvin2mev)    
-    Q2, E2, sqeset = computeSQESet(1, Q,dQ, initdos.E,dE, M, initdos.I, beta)
+    beta = 1. / (T * kelvin2mev)
+    Q2, E2, sqeset = computeSQESet(1, Q, dQ, initdos.E, dE, M, initdos.I, beta)
     # compute S(E) from SQE
     # - experiment
     # -- only need the positive part
     expsqe = sqe.copy()
-    expsqe_Epositive = expsqe[(), (-dE/2, None)].I
-    expsqeE2_Epositive = expsqe[(), (-dE/2, None)].E2
+    expsqe_Epositive = expsqe[(), (-dE / 2, None)].I
+    expsqeE2_Epositive = expsqe[(), (-dE / 2, None)].E2
     mask = expsqe_Epositive != expsqe_Epositive
     expsqe_Epositive[mask] = 0
     expsqeE2_Epositive[mask] = 0
@@ -87,31 +88,54 @@ def sqe2dos(sqe, T, Ecutoff, elastic_E_cutoff, M, initdos=None, update_weights=N
[…]
