### 2017-12-19 9ac4b98
Refs #91. added new docs to sphinx index
