### 2017-12-20 286e843
Refs #91. break index.rst into sections
