### 2017-12-20 6b1428b
Refs #91. use sphinx rtd theme
