### 2017-12-20 714e667
Refs #91. updated copyright and authors
