### 2017-12-20 fde9fa1
Refs #91. fixed some mistakes in parameter type descriptions and improved docstrings
