+++ multiphonon/forward/__init__.py
@@ -5,9 +5,9 @@
 
 import numpy as np
 def dos2sqe(dos, C_ms, sqe, T, M, Ei):
-    """calculate sqe from DOS.
+    """calculate SQE from DOS.
 
-    The computed sqe has similar props as the given (experimental) sqe.
+    The computed SQE has similar props as the given (experimental) SQE.
 
     Parameters
     ----------
@@ -29,8 +29,6 @@ def dos2sqe(dos, C_ms, sqe, T, M, Ei):
     
     Ei: float
         Incident energy (meV)
-  
-
     """
     Q, E = sqe.Q, sqe.E
     dQ = Q[1]-Q[0]; Qmin = Q[0]; Qmax=Q[-1] + dQ/2.
