### 2017-12-20 fb03169
New API documentation implemented resolving issue #91
