+++ _sources/index.txt
@@ -3,16 +3,15 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to multiphonon's documentation!
-=======================================
-
-Contents:
+API documentation for multiphonon package
+=========================================
 
 .. toctree::
    :maxdepth: 2
 
-.. automodule:: multiphonon.getdos
-		 
+.. autofunction:: multiphonon.getdos.getDOS
+.. autofunction:: multiphonon.backward.sqe2dos.sqe2dos
+		
 Indices and tables
 ==================
 
