### 2017-12-20 77e39c5
Refs #91. removed unicode characters introduced in fde9fa1
