### 2017-12-20 dea7def
Refs #97. added an example
