### 2017-12-20 1d2e96d
Refs #84. added summary to the top of API doc
