### 2017-12-20 sns-chops/multiphonon#84 @yxqd
@jochym Please check the new API doc at http://sns-chops.github.io/multiphonon/. We added API docs for  the methods intended for users. We also added a brief introduction.

We added comments to the jupyter notebook, and added [one more example](https://github.com/sns-chops/multiphonon/blob/1d2e96d98b2a42d6ab40f0f446c7750cee4e93e0/examples/getdos-Al-from-nparrs.py) to show users how to work directly from numpy arrays of I(Q,E) spectrum.
