### 2017-12-20 sns-chops/multiphonon#82 @yxqd
Thanks, @jochym. I agree a meta-package is a great idea. And being python-3 compatible is our next goal.

For installation, we removed Mantid from dependency list and updated instructions. Please give it another try. See [readme in examples](https://github.com/sns-chops/multiphonon/blob/1d2e96d98b2a42d6ab40f0f446c7750cee4e93e0/examples/README.md)
