### 2017-12-20 #440 @yxqd
Thanks @lheagy. We worked on all these tickets. I closed the DOI ticket since it is straightforward. We (@Fahima-Islam and myself) also worked on the installation and the API documentation. @jochym could you please take a look? First we would like to make sure the installation works for your system. Please give it a try and let us know if there are any problems. Thanks!
