### 2017-12-21 sns-chops/multiphonon#83 @yxqd
@jochym your environment.yml was helpful. I noticed that you have several channels in your configuration beyond what is used for multiphonon. I reproduced that in the .condarc within an ubuntu virtual machine, and it caused installation problems if mantid-framework was installed. After we try to clean up the dependency list of the mantid-framework conda package, this problem seems to go away and I can install mantid-framework without problem using all your channels + multiphonon channels. Please give it a try when you get a chance. Thanks.
