### 2018-01-02 sns-chops/multiphonon#82 @jochym
Following will be my report on installation/testing. I will amend the text as I go along the steps

1. The basic installation on the headless debian system worked fine but failed at the first example due to the lack of X11 display. I think the simple test should not require graphics if the output is graphical it should be saved to the file - this way the package can be tested in the headless environment. Also the install is still quite substantial - but I guess this cannot be easily reduced further. However, out of my curiosity: why what seems to be a computational package has so many dependencies? I would like to understand that, and be grateful if authors would comment on that point. For the install recording see: https://asciinema.org/a/Enhu0liwu5hXNwzNnT2x1lJrd
2. The example notebook required even more dependencies, as I already noted. To be fair, many of them will be already present on the typical data-science environment. On the other hand I would expect them to be present in the py3 variant and that means carrying out and maintaining separate large install just  for multiphonon. This is sub-optimal, in my opinion, and limits usefulness of the package. The size of the fully configured env with just multiphonon dependencies is 2.4GB (!), and I do not expect this data to be shared with other environments due to the python version difference. Disk space is cheap these days but this is a substantial setup which will need maintaining - and admin/user time is not getting cheaper.
3. Running the notebook led to the following error (mantid-framework is installed):
```
Converting sample data to powder I(Q,E)...
* Checking Mantid ...
RuntimeError                              Traceback (most recent call last)
/home/user/miniconda3/envs/review/lib/python2.7/site-packages/ipywe/wizard.pyc in handle_next_button_click(self, s)
     37             return
     38         self.remove()
---> 39         self.nextStep()
     40
     41     def validate(self):

/home/user/miniconda3/envs/review/lib/python2.7/site-packages/multiphonon/ui/getdos.pyc in nextStep(self)
    202             workdir=context.workdir,
    203         )
--> 204         for msg in r2i:print msg
    205         context.iqe_h5, context.mtiqe_h5, context.Qaxis, context.Eaxis = msg
    206         print "Done."

/home/user/miniconda3/envs/review/lib/python2.7/site-packages/multiphonon/getdos.pyc in reduce2iqe(sample_nxs, Emin, Emax, dE, Qmin, Qmax, dQ, mt_nxs, iqe_nxs, iqe_h5, workdir)
    123     Qaxis = _normalize_axis_setting(Qmin, Qmax, dQ)
    124     yield "Converting sample data to powder I(Q,E)..."
--> 125     raw2iqe(sample_nxs, iqe_nxs, iqe_h5, Eaxis, Qaxis, type='sample')
    126     if mt_nxs is not None:
    127         _tomtpath = lambda p: os.path.join(

/home/user/miniconda3/envs/review/lib/python2.7/site-packages/multiphonon/getdos.pyc in raw2iqe(eventnxs, iqe_nxs, iqe_h5, Eaxis, Qaxis, type)
    181         os.remove(iqe_nxs); os.remove(iqe_h5)
    182     #
--> 183     from .redutils import reduce, extract_iqe
    184     Emin, Emax, dE = Eaxis
    185     Emin-=dE/2; Emax-=dE/2 # mantid algo use bin boundaries

/home/user/miniconda3/envs/review/lib/python2.7/site-packages/multiphonon/redutils.py in <module>()
     26     return
     27 if not mantid_checked:
---> 28     _checkMantid()
     29
     30

/home/user/miniconda3/envs/review/lib/python2.7/site-packages/multiphonon/redutils.py in _checkMantid()
     20     sp.call(shlex.split("python -c 'import mantid'"), stdout=sp.PIPE, stderr=sp.PIPE) # sometimes mantid import for the first time may fail
     21     if sp.call(shlex.split("python -c 'import mantid'")):
---> 22         raise RuntimeError("Please install mantid")
     23     global mantid_checked
     24     mantid_checked = True

RuntimeError: Please install mantid
```
4. Tests:
  - The "no mantid, no ipwe" test run to completion, however with the two warnings:
```warnings summary:
ms_TestCase.py::TestCase::test1
  /home/user/miniconda3/envs/review/lib/python2.7/site-packages/histogram/hdf/Loader.py:129: FutureWarning: elementwise comparison failed; returning scalar instead, but in the future will perform elementwise comparison
    if 'storage' in list(dataGroup): # this uses the 'storage' convention

backward/singlephonon_sqe2dos_TestCase.py::TestCase::test1c
  /home/user/multiphonon/multiphonon/backward/singlephonon_sqe2dos.py:158: UserWarning: Scaling factor to combine DOSes calculated is not stable: 22.1354562834 (using continuous criteria) vs 17.7865062045 (using area criteria)
  You may want to consider adjusting the parameters such as E range (Emax more specifically)
    scale1, scale2)
```
The last warning seems troubling - is that expected? In the test case?
  - The full test set failed with 5 failing tests and `RuntimeError: Please install mantid` message. As I stated above mantid-framework is installed.

All above tests were performed in the clean, fresh install of debian 9.3 on the headless virtual machine.
The tests on workstation install will follow in the next post.
