### 2018-01-02 sns-chops/multiphonon#82 @jochym
Tests with desktop system:

1. Install went the same - no problems this time
2. All tests executed, eventually.
3. The tests seem to be rather memory-hungry. I was unable to complete all of them on the system with 4GB of RAM. They finally completed when I increased memory to 8GB but 5GB would probably be enough. It is rather large. Is this an inherent property of the code or could authors do something about it? If nothing could be done the docs should clearly state this requirement - for that type of calculation I did not expect such requirements.
3. The example notebook run to completion and produced expected results.
4. I am not too fond of "wizards" used in the notebook. They are not very clear and break a natural workflow of the document. I would rather have clearly commented dictionaries with parameters and functions you can just run and wait for the result. Note that this type of the interface makes "run all cells" command useless. The same goes for expected in the future notebook import ability of the jupyter system. If the authors want to build an interactive dashboard/application they should do that. I my opinion it is better when the notebook keeps the document metafor. The wizards bring unnecessary "moving parts" to the example.
5. The "instability" warnings kept showing up in the test results. I think authors should address this concern.
6. The full system (with jupyter interface) cannot be used on the headless server - it seems to me. I suspect authors cannot do anything about it, as I think this limitation comes from mantid component. At the same time this is not a real limitation since the process is *not* actually using any X11 graphics - it just expects to have one active. This limits the usability of the system. E.g. I could not use it on my production system as this is purely headless, networked jupyterhub install on the cloud service. I would like to have authors address/comment this point as well.
