### 2018-01-02 sns-chops/multiphonon#82 @jochym
@yxqd @lheagy : Final remarks (install)
----------------------------
The basic problems are resolved. The summary of remaining install issues:
1. Problems on headless system with non-helpful error messages. Either solve the problem or explain it in the docs. It will be very confusing for inexperienced user.
2. Memory requirements should be explained/documented or reduced
3. Address instability warnings in the tests
4. At least comment on the issues noted in my two above posts.
