### 2018-01-02 sns-chops/multiphonon#83 @jochym
As I noted in the https://github.com/sns-chops/multiphonon/issues/82 the clean install in the freshly installed debian desktop machine with clean conda worked as expected. Note however that this is not a typical setup and you cannot expect that users will go to such lengths to run the software. Additionally as
 the install on my standard system still fails to work.
When the software worked, I was able to reproduce expected results. But the examples are still fairly limited and tersly documented.

In summary, this time I was able to at least verify that the expected results can be reproduced. I still think that they are too sparsely documented but they exist and work.
