### 2018-01-02 sns-chops/multiphonon#84 @jochym
- The api docs are greatly expanded and much better now.
- Can you @yxqd declare that every function in your public API is documented? This is important - there should be no undocumented parts of the API which are required for the functionality of the software. I am fine with undocumented internal interfaces but strongly advise against keeping substantial APIs (external or _internal_) undocumented.
- The additional example is welcomed but should be mentioned in the readme and explained in the docs
- I still cannot find any functional documentation of the package. I think that this is fairly important - the beginner user has no clear indications where to start. There is no "principles of operation" text and no explanation of rather cryptic and non-standard steps in the notebook. The learning curve for someone not familiar with mandit will be rather steep. Such docs should be included in the RTD documents.
