### 2018-01-02 #440 @jochym
@lheagy see my comments in the review issues.
