### 2018-01-02 sns-chops/multiphonon#82 @yxqd
Thanks @jochym! I will try to add a new issue on making things working on a headless system (we did not expect this kind of usage but I agree this might come up in the future). Regarding memory requirements: the mantid reduction requires quite a bit of memory. I will try to add comments to documentation about this. It looks like mantid installation using conda is still causing problems on your standard system. Most of our current users are using the SNS analysis clusters (redhat) in which mantid conda installation does work. Fixing the mantid-conda installation is beyond the scope of this paper I think. And I agree leaving it out of the dependency list is the right call. Sorry I am not back to work yet and will add more details soon, most likely tomorrow.
