### 2018-01-02 sns-chops/multiphonon#82 @jochym
Do not worry too much about my system - you are not responsible for such things. My system is not standard and some brakages may be expected. It is important that the soft work on standard released distros - and it does at least on last debian. Really important point is the headless systems. This is quite common use case (cloud services!) and I believe there is no _real_ requirement for X11 in mantid. Maybe it can be just reconfigured to use different backend at runtime or something? It does not actually _use_ X11 when it runs in the notebook.
