### 2018-01-03 sns-chops/multiphonon#100 @yxqd
getdos-Al example script: make it work for headless system
See #82
