### 2018-01-03 sns-chops/multiphonon#82 @yxqd
Thanks @jochym. Glad that the software works for standard debian linux distributions. We have only tested ubuntu/fedora/redhat before.

Regarding headless systems, it turns out the core functionality of multiphonon works. The example script failed at a headless system only because it generates a plot at the end for users to interact with. I just created #100 to address this.

The Mantid software was started about 10 years ago and at that time it was not really intended to be a python library. Instead, the focus was mantidplot, the GUI. Only very recently it is pushed for being a stand-alone python library. So I think the task of making sure mantid works for a variety of headless systems is a long battle for the future. For example, I tried installing mantid-framework conda package on a miniconda docker image (which derives from debian), and "import mantid" only works after "freeglut3-dev" is installed using apt-get.
