### 2018-01-03 sns-chops/multiphonon#101 @yxqd
Improve documentation
See #82, #83, #84

* [x] Installation: comment on mantid conda installation limitations
* [x] add hints on memory usage
* [x] add a section on stitching
* [x] add a new notebook without UI
