### 2018-01-03 a89cf94
update examples/README.md
explain the new script getdos-Al-from-nparrs.py
