### 2018-01-03 sns-chops/multiphonon#82 @jochym
1. Regarding mantid - this is fine. You are not submitting mantid for review. You are not responsible for its deficiencies. But they influence your software in quite cryptic way, and at the beginning of the learning curve. Thus I think that these problems should be at least mentioned (better still explained) in the installation docs.
2. This text about the warning should be included in the docs - it will help users gain confidence in the package.
