### 2018-01-03 sns-chops/multiphonon#84 @yxqd
Thanks @jochym.

* Great!
* I just went through the API doc and yes every public function is documented. There are some functions that may be worth considered to be made public in the future, but I don't think they are important for this version of the software.
* Added explanation: a89cf94
* That particular notebook is intended for users who are not programmers and that is why we provide some small UI widgets in the notebook to guide the users through the procedure, and it is hiding a lot of the technical details. I will add another notebook to make it clear what is going on behind the UI widgets.
