### 2018-01-03 sns-chops/multiphonon#83 @yxqd
Thanks @jochym. We are glad that it worked for a standard debian distribution you tried out. Please see my posts in #82 for details. And some documentation improvements are coming: #101
