### 2018-01-03 sns-chops/multiphonon#84 @jochym
The additional notebook will be a very welcomed addition.
