### 2018-01-04 e669469
Refs #100. check for DISPLAY and do not plot if it is not set
