+++ examples/getdos-Al.py
@@ -5,7 +5,9 @@ All this script does is to convert I(Q,E) to Density of States
 by performing multiphonon correction.
 """
 import histogram.hdf as hh, os, numpy as np
-from matplotlib import pyplot as plt
+headless = 'DISPLAY' not in os.environ
+if not headless:
+    from matplotlib import pyplot as plt
 # load data
 iqehist = hh.load("data/Al-iqe.h5") 
 # interpolate data
@@ -19,8 +21,9 @@ iterdos = sqe2dos.sqe2dos(
     newiqe, T=300, Ecutoff=50., 
     elastic_E_cutoff=(-10., 7), M=26.98,
     C_ms=0.2, Ei=80., workdir='work-Al')
-# process and plot
-for i, dos in enumerate(iterdos):
-    plt.plot(dos.E, dos.I, label='%d' % i)
-plt.legend()
-plt.show()
+if not headless:
+    # process and plot
+    for i, dos in enumerate(iterdos):
+        plt.plot(dos.E, dos.I, label='%d' % i)
+    plt.legend()
+    plt.show()
