+++ examples/getdos-Al-from-nparrs.py
@@ -3,9 +3,14 @@ This is a simple example where I(Q,E) data is already in the form
 of a histogram, reduced from the raw experimental data.
 All this script does is to convert I(Q,E) to Density of States
 by performing multiphonon correction.
+The different between this script and the "getdos-Al.py" script
+is that here the histogram needs to be constructed from numpy
+arrays of Q, E axes and intensity and squared errorbar matrices.
 """
 import histogram.hdf as hh, histogram as H, os, numpy as np
-from matplotlib import pyplot as plt
+headless = 'DISPLAY' not in os.environ
+if not headless:
+    from matplotlib import pyplot as plt
 # load data
 Q = np.load("data/Al-iqe-Q.npy") 
 E = np.load("data/Al-iqe-E.npy") 
@@ -24,8 +29,9 @@ iterdos = sqe2dos.sqe2dos(
     newiqe, T=300, Ecutoff=50., 
     elastic_E_cutoff=(-10., 7), M=26.98,
     C_ms=0.2, Ei=80., workdir='work-Al-from-nparrs')
-# process and plot
-for i, dos in enumerate(iterdos):
-    plt.plot(dos.E, dos.I, label='%d' % i)
-plt.legend()
-plt.show()
+if not headless:
+    # process and plot
+    for i, dos in enumerate(iterdos):
+        plt.plot(dos.E, dos.I, label='%d' % i)
+    plt.legend()
+    plt.show()
