### 2018-01-04 ea83f1e
Refs #100. added example scripts to travis CI
