### 2018-01-04 19ae2f4
Refs #101. moved installation and usage from README.md to docs/index.rst
