### 2018-01-04 f1fc635
Refs #101. added comments on mantid-framework conda installation
