### 2018-01-04 190eea7
Refs #101. comments on system requirements
