### 2018-01-04 8c9872b
Refs #101. added new example notebook without UI. added api doc for getdos.reduce2iqe
