+++ multiphonon/getdos.py
@@ -109,6 +109,43 @@ def getDOS(sample_nxs, mt_nxs=None, mt_fraction=0.9, const_bg_fraction=0.,
 
 
 def reduce2iqe(sample_nxs, Emin,Emax,dE, Qmin,Qmax,dQ, mt_nxs=None, iqe_nxs='iqe.nxs', iqe_h5='iqe.h5', workdir='work'):
+    """Reduce sample and (optionally) empty can nxs files and generate I(Q,E) histograms
+
+    Parameters
+    ----------
+    sample_nxs : str
+        Sample Nexus file
+
+    Emin : float
+        Energy tranfer axis minimum
+
+    Emax : float
+        Energy tranfer axis maximum
+
+    dE : float
+        Energy tranfer axis step size
+
+    Qmin : float
+        Momentum tranfer axis minimum
+
+    Qmax : float
+        Momentum tranfer axis maximum
+
+    dQ : float
+        Momentum tranfer axis step size
+
+    mt_nxs : str
+        Empty can Nexus file
+
+    iqe_nxs : str
+        nexus filename of reduced I(Q,E). temporary file.
+
+    iqe_h5: str
+        output histogram filename of reduced I(Q,E)
+
+    workdir: str
+        path to working directory
+    """
     # prepare paths
     if not os.path.exists(workdir):
         os.makedirs(workdir)
