### 2018-01-04 1d15822
Refs #101. no-UI example notebook: clear output
