### 2018-01-05 53e3ba0
Refs #101. added two sections to usage.rst
