+++ docs/usage.rst
@@ -24,12 +24,89 @@ For any user, GetDOS can be performed with a local installation of multiphonon.
 Examples and instructions can be found `here <https://github.com/sns-chops/multiphonon/tree/master/examples>`_
 
 
+General principle of operation
+------------------------------
 
-Multiple-Ei and Stitching
--------------------------
+The main functionality of this package is to convert S(Q,E) spectrum
 
-About the last warning in the test: it is expected since we are still experimenting on this new feature for "stitching"
-together phonon DOSes measured at different neutron incident energies. Right now researchers more or less do the stitching
-manually and somewhat arbitrarily, and this software provides two routines for stitching.
-The warning is to alert the users that two methods of stitching were used but they disagree
-somewhat in the calculation of scaling factor for stitching. The disagreement is actually acceptable in this case.
+.. image:: images/V-sqe.png
+   :width: 400px
+
+to a phonon DOS spectrum
+
+.. image:: images/V-dos.png
+   :width: 400px
+
+The best way to learn how to perform such a conversion is to follow one of the notebooks
+or scripts in `examples <https://github.com/sns-chops/multiphonon/tree/master/examples>`_
+	   
+The processing is an iterative procedure, and in this plot you can see the DOS
+is getting better in each iteration:
+
+.. image:: images/V-DOS-iters.png
+   :width: 400px
+
+
+Working with datasets from multiple incident energies
+-----------------------------------------------------
+
+This is an advanced topic.
+
+Direct geometry spectrometers are known for their peculiar resolution variation:
+the resolution is the best when the energy transfer is close to the incident energy
+(the neutron final energy is near zero), and then it gets worse (broader) and worse
+as the energy transfer gets smaller.
+As a result, for a material with a wide range of phonon energies, the resolution of DGS instrument
+at lower energy enegies is not satisfactory.
+One way to alleviate this problem is to measure the phonon spectrum using
+multiple incident energies.
+
+This is best explained with an example.
+For a system with phonon energies up to 250meV, we
+could measure the spectrum at two incident energies at ARCS:
+300meV, and 130meV.
+
+Here is a plot of the density of states obtained from the :math:`E_i=300` meV dataset:
+
+.. image:: images/graphite-DOS-Ei_300.png
+   :width: 400px
+
+This data covers the full phonon DOS spectrum, but the resolution at lower energies are not optimal.
+	   
+Another dataset was meassured at :math:`E_i=130` meV. This is the measured S(Q,E):
+
+.. image:: images/graphite-SQE-Ei_130.png
+   :width: 400px
+
+The DOS measured in this dataset cannot exceeds 130meV, but is of better energy resolution:
+
+.. image:: images/graphite-DOS-Ei_130.png
+   :width: 400px
+
+The two DOSes should be "stiched" together to obtain one DOS curve.
+This is done by updating the low :math:`E` portion of the DOS from the higher :math:`E_i` measurement
+with the partial DOS obtained from the lower :math:`E_i` measurement.
+
+Two algorithms are available to calculate the scale factor between the two DOSes.
+One algorithm updates the lower :math:`E` portion of the DOS by keeping the area of the updated portion intact
+(area rule),
+the other algorithm updates the lower :math:`E` portion of the DOS
+by keeping the DOS value at the stiching point the same as the original DOS (continuity rule).
+Ideally these two algorithms should give the same scaling factors, but they always disagree
+somewhat.
+A good rule of thumb is to choose the stitching point (Ecutoff) at a relatively flat portion of the DOS
+curve.
+
+When the two algorithms disagree in calculating the scaling factor, you may see a warning message like this::
+
[…]
