+++ examples/README.md
@@ -42,3 +42,10 @@ Start a jupyter server
 And then open an example notebook and follow instructions
 
 **Note:** For an example notebook, look for a pdf with same name. It should contain expected outputs and plots.
+
+For users who are not familar with python and more comfortable with GUI, please start with
+[the vanadium example with UI widgets](getdos2-V_Ei120meV.ipynb).
+
+For users who are comfortable with python and jupyter notebook, please start with
+[the vanadium example without UI widgts][getdos2-V_Ei120meV-noUI.ipynb).
+
