### 2018-01-05 9886665
Refs #83, #101. added a bit of documentation to the example scripts
