### 2018-01-05 9f3e301
Refs #83, #101. examples/README: minor wording change
