### 2018-01-05 adf13f3
Refs #101. gh-pages branch: minor wording update
