### 2018-01-05 6cbecf4
Refs #101. minor wording change to installation and usage docs
