### 2018-01-05 ec6c84e
Refs #104. update _version.py
