### 2018-01-05 46f18f3
Refs #104. updated conda-recipes/meta.yaml
