+++ conda-recipe/meta.yaml
@@ -1,6 +1,6 @@
 package:
   name: multiphonon
-  version: "0.1.2a1"
+  version: "0.1.2a2"
 
 source:
   git_rev: master
