### 2018-01-07 #440 @lheagy
Thanks @jochym for all of your feedback! @yxqd, please keep us posted on the progress of the issues, as it looks like
- sns-chops/multiphonon#82
- sns-chops/multiphonon#83
- sns-chops/multiphonon#84
are still in progress
