### 2018-01-08 #440 @yxqd
Thanks @lheagy. We tried to address them in sns-chops/multiphonon#101 and sns-chops/multiphonon#100. @Fahima-Islam is giving it a final check and then we will ask @jochym for another review.
