### 2018-01-08 e018561
minor fixes of docs
