+++ docs/conf.py
@@ -113,6 +113,7 @@ todo_include_todos = False
 # a list of builtin themes.
 html_theme = 'alabaster'
 html_theme = 'sphinx_rtd_theme'
+# html_theme = 'default'
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
