### 2018-01-08 sns-chops/multiphonon#82 @yxqd
@jochym @lheagy We made the following improvements:

* Moved the majority of the installation and usage documentation from README.md to the documentation site http://sns-chops.github.io/multiphonon/ and expand them:
  - explained [limitations of mantid-framework conda installation](http://sns-chops.github.io/multiphonon/installation.html#notes-on-mantid-conda-installation)
  - added a section on [system requirements](http://sns-chops.github.io/multiphonon/installation.html#system-requirements)
  - added a section on "stitching": http://sns-chops.github.io/multiphonon/usage.html#working-with-datasets-from-multiple-incident-energies
* Add new example notebook and script:
  - [Example notebook without UI widgets](https://github.com/sns-chops/multiphonon/blob/a65dcf0555e4410562ffc14eba5a302812d8d01c/examples/getdos2-V_Ei120meV-noUI.ipynb)
  - [Example script for stitching](https://github.com/sns-chops/multiphonon/blob/a65dcf0555e4410562ffc14eba5a302812d8d01c/examples/getdos2-graphite_Ei130meV-noUI.py)
