### 2018-01-08 #440 @yxqd
@lheagy @jochym Please see updates at sns-chops/multiphonon#82
