### 2018-01-11 sns-chops/multiphonon#82 @lheagy
Thanks @yxqd, @jochym, in your opinion, can we close this issue?
