### 2018-01-28 sns-chops/multiphonon#82 @jochym
I like what you did with the docs and install description. I would include the actual command-by-command tutorial in the docs, but it is present in the repo and linked from the docs so this is only advice not a complaint. Please add that it will probably not work on the headless system without special config - it is better to not feed false hopes of the users. I like what you did with the examples as well.
I would suggest considering producing mybinder (https://mybinder.org/) repo for the package. It is quite easy for the conda installable package, and it would help users to check if they have the install correctly working. But this is only a suggestion, and not a part of the review.
I had to add `import mantid` to the notebook for it to work on my system. It took several minutes maybe warn the users to expect long run times and not panic?
It will be better to add this even if on some systems I tested it was not required - it cannot hurt.
And maybe change download part to not run each time if the data is already there ?
Otherwise I think this issue can be closed after this small things are fixed.
