### 2018-01-28 #440 @jochym
@lheagy, @yxqd : Following my comments in https://github.com/sns-chops/multiphonon/issues/82 I think that after addressing  the last minor issues mentioned there the paper can be accepted to be published in JOSS.
