### 2018-01-29 5fe2705
Refs #82. Added notes on headless systems
