### 2018-01-29 47a4e57
Refs #82. explicitly import mantid in example notebooks
