### 2018-01-29 sns-chops/multiphonon#82 @yxqd
Thanks @jochym. I put in some minor fixes as you suggested and created tickets for some other minor points to work on later. I will close this ticket and #83, #84 now. Please let me know if there are any other problems.
