### 2018-01-29 #440 @yxqd
@jochym Thanks again for the review. @jochym @lheagy I addressed some minor fixes in sns-chops/multiphonon#82 and created a couple issues for @jochym 's suggestions to work on later (beyond this review). Please check. Thanks!
