### 2018-01-29 #440 @lheagy
Wonderful, thanks for all of your work @jochym and @yxqd.

I have one last request, @jochym, could you take a look at the checklist at the top of this issues and if all items have been addressed, check them off? Thank you!

@bjmorgan: are there any other comments you have regarding this submission?
