### 2018-01-29 #440 @arfon
@jochym - can you visit this page: https://github.com/openjournals/joss-reviews/invitations and accept the invite at the top of the page? This should allow you to modify the checklist.
